Cleanup common.cpp
