+++ Samples/WindowsML/cpp/CppConsoleDesktop.GenAI/timing.h
